buildkite: rename files to fix lint issues
